+++ Classes/SMLSnippetsController.m
@@ -199,7 +199,7 @@ static id sharedInstance = nil;
 }
 
 
-- (void)openPanelDidEnd:(NSOpenPanel *)panel returnCode:(int)returnCode  contextInfo:(void  *)contextInfo
+- (void)openPanelDidEnd:(NSOpenPanel *)panel returnCode:(NSInteger)returnCode  contextInfo:(void  *)contextInfo
 {
 	if (returnCode == NSOKButton) {
 		[self performSnippetsImportWithPath:[panel filename]];
