+++ ICU/NSStringICUAdditions.m
@@ -46,7 +46,7 @@ typedef struct URegularExpression URegularExpression;
 	NSMutableArray *foundGroups = [NSMutableArray array];
 	
 	[matcher findFromIndex:0];	
-	int i;
+	NSInteger i;
 	for(i=0;i<=[matcher numberOfGroups];i++)
 		[foundGroups addObject:[matcher groupAtIndex:i]];
 
@@ -80,13 +80,13 @@ typedef struct URegularExpression URegularExpression;
 	
 	// for some reason, the null-terminator doesn't always show up at the right place and this
 	// causes extra characters to be created in the unicode string.  We remove them here by force.
-	unsigned int len = [self length];
+	NSUInteger len = [self length];
 	ret[len] = '\0';
 	return ret;	
 }
 
 -(void *)copyUTF16String {
-	unsigned int length = [self length];
+	NSUInteger length = [self length];
 	UChar *utf16String = malloc((length+1)*sizeof(UChar));
 	[self getCharacters: utf16String];
 	utf16String[length] = 0;
