+++ PSMTabBar/PSMTabBarCell.h
@@ -16,11 +16,11 @@
     // sizing
     NSRect              _frame;
     NSSize              _stringSize;
-    int                 _currentStep;
+    NSInteger                 _currentStep;
     BOOL                _isPlaceholder;
     
     // state
-    int                 _tabState;
+    NSInteger                 _tabState;
     NSTrackingRectTag   _closeButtonTrackingTag;    // left side tracking, if dragging
     NSTrackingRectTag   _cellTrackingTag;           // right side tracking, if dragging
     BOOL                _closeButtonOver;
@@ -30,7 +30,7 @@
     BOOL                _hasCloseButton;
     BOOL                _isCloseButtonSuppressed;
     BOOL                _hasIcon;
-    int                 _count;
+    NSInteger                 _count;
 }
 
 // creation/destruction
@@ -44,14 +44,14 @@
 - (void)setCloseButtonTrackingTag:(NSTrackingRectTag)tag;
 - (NSTrackingRectTag)cellTrackingTag;
 - (void)setCellTrackingTag:(NSTrackingRectTag)tag;
-- (float)width;
+- (CGFloat)width;
 - (NSRect)frame;
 - (void)setFrame:(NSRect)rect;
 - (void)setStringValue:(NSString *)aString;
 - (NSSize)stringSize;
 - (NSAttributedString *)attributedStringValue;
-- (int)tabState;
-- (void)setTabState:(int)state;
+- (NSInteger)tabState;
+- (void)setTabState:(NSInteger)state;
 - (NSProgressIndicator *)indicator;
 - (BOOL)isInOverflowMenu;
 - (void)setIsInOverflowMenu:(BOOL)value;
@@ -65,18 +65,18 @@
 - (BOOL)isCloseButtonSuppressed;
 - (BOOL)hasIcon;
 - (void)setHasIcon:(BOOL)value;
-- (int)count;
-- (void)setCount:(int)value;
+- (NSInteger)count;
+- (void)setCount:(NSInteger)value;
 - (BOOL)isPlaceholder;
 - (void)setIsPlaceholder:(BOOL)value;
-- (int)currentStep;
-- (void)setCurrentStep:(int)value;
+- (NSInteger)currentStep;
+- (void)setCurrentStep:(NSInteger)value;
 
 // component attributes
 - (NSRect)indicatorRectForFrame:(NSRect)cellFrame;
 - (NSRect)closeButtonRectForFrame:(NSRect)cellFrame;
-- (float)minimumWidthOfCell;
-- (float)desiredWidthOfCell;
+- (CGFloat)minimumWidthOfCell;
+- (CGFloat)desiredWidthOfCell;
 
 // drawing
 - (void)drawWithFrame:(NSRect)cellFrame inView:(NSView *)controlView;
